# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.LoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse:
    properties:
      data:
        items:
          type: integer
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioCreateDTO:
    properties:
      email:
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
      preferencias:
        items:
          type: integer
        type: array
      senha:
        minLength: 6
        type: string
    required:
    - email
    - nome
    - senha
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioResponseDTO:
    properties:
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      preferencias:
        items:
          type: integer
        type: array
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioUpdateDTO:
    properties:
      email:
        type: string
      nome:
        maxLength: 100
        minLength: 3
        type: string
      preferencias:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Cultivo app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cultivo API
  version: "1.0"
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Recebe credenciais de usuário (email e senha) e retorna um token
        JWT para acesso autenticado.
      parameters:
      - description: Credenciais do Usuário
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autentica um usuário e retorna um token JWT
      tags:
      - usuario
...
  /api/v1/usuarios:
    get:
      description: Retorna uma lista paginada de todos os usuários cadastrados
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limite de itens por página (padrão: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista todos os usuários com paginação
      tags:
      - usuario
    post:
      consumes:
      - application/json
      description: Cria um novo usuário com os dados fornecidos
      parameters:
      - description: Dados do Usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo usuário
      tags:
      - usuario
  /api/v1/usuarios/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um usuário existente
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta um usuário
      tags:
      - usuario
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um usuário específico
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca um usuário por ID
      tags:
      - usuario
    put:
      consumes:
      - application/json
      description: Atualiza um usuário existente com os dados fornecidos
      parameters:
      - description: ID do Usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do Usuário para atualização
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UsuarioResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um usuário
      tags:
      - usuario
...
swagger: "2.0"